patient-records-secure/requirements.txt

Flask==3.0.3
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-SQLAlchemy==3.1.1
email-validator==2.2.0
python-dotenv==1.0.1
cryptography==43.0.1
passlib==1.7.4
